requirements.txt
fastapi
uvicorn
python-multipart
pydantic-settings
groq
edge_tts
mega.py
pydub
soundfile
tenacity
